fix(session): preserve complete transcripts
