+++ tests/agent/test_consolidator.py
@@ -363,7 +363,7 @@ class TestConsolidatorTokenBudget:
             await consolidator.maybe_consolidate_by_tokens(session, runtime=runtime)
 
     async def test_estimate_uses_full_unconsolidated_tail(self, consolidator, runtime):
-        """Consolidation pressure must see messages hidden by the replay window."""
+        """Consolidation pressure must account for the full unarchived tail."""
         session = Session(key="test:full-tail")
         for i in range(160):
             session.add_message("user", f"msg-{i}")
@@ -399,107 +399,6 @@ class TestConsolidatorTokenBudget:
 
         assert len(captured["history"]) == 8
         assert captured["history"][0]["content"] == "msg-2"
-
-    async def test_replay_window_overflow_is_archived_even_under_token_budget(
-        self,
-        consolidator,
-        runtime,
-    ):
-        """Old messages that cannot be replayed should be materialized first."""
-        consolidator._SAFETY_BUFFER = 0
-        session = Session(key="test:replay-overflow")
-        session.provider_state = _provider_state()
-        for i in range(10):
-            session.add_message("user", f"u{i}")
-            session.add_message("assistant", f"a{i}")
-
-        consolidator.sessions._session_cache[session.key] = session
-        consolidator.estimate_session_prompt_tokens = MagicMock(return_value=(100, "tiktoken"))
-        consolidator.archive_session = AsyncMock(return_value="old conversation summary")
-
-        await consolidator.maybe_consolidate_by_tokens(
-            session,
-            runtime=runtime,
-            replay_max_messages=6,
-        )
-
-        archive_end = consolidator.archive_session.await_args.kwargs["archive_end"]
-        archived_chunk = session.messages[:archive_end]
-        assert archived_chunk[0]["content"] == "u0"
-        assert archived_chunk[-1]["content"] == "a6"
-        assert session.last_consolidated == 14
-        assert session.metadata["_last_summary"]["text"] == "old conversation summary"
-        assert session.provider_state is None
-        consolidator.sessions.save.assert_called()
-
-    async def test_replay_window_overflow_extends_to_long_recent_user_turn(
-        self,
-        consolidator,
-        runtime,
-    ):
-        """Replay-window consolidation must not cut into the latest user turn."""
-        session = Session(key="test:replay-tool-boundary")
-        session.add_message("user", "old")
-        session.add_message("assistant", "old answer")
-        session.add_message("user", "record this")
-        for i in range(4):
-            session.messages.extend(_tool_round(f"call-{i}"))
-        session.add_message("assistant", "final answer")
-
-        consolidator.sessions._session_cache[session.key] = session
-        consolidator.estimate_session_prompt_tokens = MagicMock(return_value=(100, "tiktoken"))
-        consolidator.archive_session = AsyncMock(return_value="tool turn summary")
-
-        await consolidator.maybe_consolidate_by_tokens(
-            session,
-            runtime=runtime,
-            replay_max_messages=4,
-        )
-
-        archive_end = consolidator.archive_session.await_args.kwargs["archive_end"]
-        archived_chunk = session.messages[:archive_end]
-        assert [m["content"] for m in archived_chunk] == ["old", "old answer"]
-        assert session.last_consolidated == 2
-
-        history = session.get_history(max_messages=4, extend_to_user=True)
-        assert len(history) > 4
-        assert history[0]["content"] == "record this"
-        assert history[-1]["content"] == "final answer"
-
-    async def test_replay_window_overflow_uses_newer_user_inside_window(
-        self,
-        consolidator,
-        runtime,
-    ):
-        """Do not extend to an older long turn when the hard window has a newer user."""
-        session = Session(key="test:replay-newer-user")
-        session.add_message("user", "old")
-        session.add_message("assistant", "old answer")
-        session.add_message("user", "long older turn")
-        for i in range(8):
-            session.messages.extend(_tool_round(f"older-{i}"))
-        session.add_message("assistant", "older final")
-        session.add_message("user", "new question")
-        session.add_message("assistant", "new answer")
-
-        consolidator.sessions._session_cache[session.key] = session
-        consolidator.estimate_session_prompt_tokens = MagicMock(return_value=(100, "tiktoken"))
-        consolidator.archive_session = AsyncMock(return_value="older turn summary")
-
-        await consolidator.maybe_consolidate_by_tokens(
-            session,
-            runtime=runtime,
-            replay_max_messages=6,
-        )
-
-        archive_end = consolidator.archive_session.await_args.kwargs["archive_end"]
-        archived_chunk = session.messages[:archive_end]
-        assert archived_chunk[2]["content"] == "long older turn"
-        assert archived_chunk[-1]["content"] == "older final"
-        assert session.last_consolidated == len(session.messages) - 2
-
-        history = session.get_history(max_messages=6, extend_to_user=True)
-        assert [m["content"] for m in history] == ["new question", "new answer"]
 
     async def test_token_overflow_appends_prompt_to_replay_prefix(
         self,
